+++ packages/lib/src/games.ts
@@ -30,7 +30,12 @@ export const games: Array<Game> = [
       games: [
         {
           title: "Blue Protocol: Star Resonance",
-          processNames: ["Star.exe", "BPSR_STEAM.exe", "BPSR.exe"],
+          processNames: [
+            "Star.exe",
+            "BPSR_STEAM.exe",
+            "BPSR.exe",
+            "BPSR_EPIC.exe",
+          ],
         },
       ],
       defaultHotkeys: {